# api/openapi-lock.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "org-quicko-cliq",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{base_url}}"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/programs/{program_id}/promoters/{promoter_id}/links": {
      "post": {
        "tags": [
          "Link"
        ],
        "summary": "Create link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "link"
                  },
                  "link_id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "url": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "name",
                  "url"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "name": "",
                    "url": "quicko.com/it/plans"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully created link\",\r\n    \"data\": {\r\n        \"link_id\": \"\",\r\n        \"name\": \"\",\r\n        \"created_at\": \"\",\r\n        \"updated_at\": \"\"\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Link"
        ],
        "summary": "Get all links",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got all links\",\r\n    \"data\": {\r\n        \"links\": [\r\n            {\r\n                \"link_id\": \"\",\r\n                \"name\": \"\",\r\n                \"created_at\": \"\",\r\n                \"updated_at\": \"\"\r\n                \r\n            },\r\n             {\r\n                \"link_id\": \"\",\r\n                \"name\": \"\",\r\n                \"created_at\": \"\",\r\n                \"updated_at\": \"\"\r\n                \r\n            },\r\n             {\r\n                \"link_id\": \"\",\r\n                \"name\": \"\",\r\n                \"created_at\": \"\",\r\n                \"updated_at\": \"\"\r\n                \r\n            }\r\n        ]\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/links/{link_id}": {
      "delete": {
        "tags": [
          "Link"
        ],
        "summary": "Delete a link",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{link_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully removed Link\",\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
...
  }
}